Fix PUBLIC_URL path in docker production
After applying this commit, the env PUBLIC_URL
will only be applied to resource paths if it is
not empty and not the dot character.

This change is necessary, as env.js will copy the
dot of the homepage into the PUBLIC_URL,
while CRA would actually strip it. react-router
will fail to correctly link the internal paths with
the a leading dot.